% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{add_title_header}
\alias{add_title_header}
\title{Add a title in headers}
\usage{
add_title_header(x, title, font_size = 12)
}
\arguments{
\item{x}{a \code{flextable} object}

\item{title}{a string containing the title of the table}

\item{font_size}{integer to use for font size of the title}
}
\description{
Add a title in the flextable's header part. It can
be inserted at the top or the bottom of header part.
}
\details{
An horizontal and a vertical merge are performed automatically.
Use \code{\link{merge_none}} to drop theses merging instructions.
}
\examples{
ft <- flextable::flextable( utils::head( iris ),
  col_keys = c("Species", "Sepal.Length", "Petal.Length", "Sepal.Width", "Petal.Width") )
ft_wtitle <- add_title_header(ft, "Iris data set")
}
