% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{ft_theme}
\alias{ft_theme}
\title{Apply theme to a flextable}
\usage{
ft_theme(x, add_w = 0.1, add_h = 0.05, add_h_header = 0.05)
}
\arguments{
\item{x}{a flextable object}

\item{add_w}{extra width to add in inches}

\item{add_h}{extra height to add in inches}

\item{add_h_header}{extra header height to add in inches}
}
\description{
Apply theme to a flextable
}
\examples{
ft <- flextable::flextable(iris)
ft <- ft_theme(ft)
}
