% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{gg_pop_theme}
\alias{gg_pop_theme}
\title{Custom ggplot2 theme}
\usage{
gg_pop_theme(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{default font size}

\item{base_family}{font family to use. Default is "sans".}
}
\description{
Use ggplot2's theming facility to set defaults
}
\details{
This function sets the default look (colors, linetypes, etc.) of ggplot2 to
match color scheme.  Exhibits not made with ggplot2 will not
be affected.
}
\examples{

gg_pop_theme()

}
\seealso{
\code{\link{theme_bw}}, \code{\link{theme_set}}
}
