% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\docType{data}
\name{pop_colors}
\alias{pop_colors}
\title{Default Color Scheme}
\format{An object of class \code{list} of length 19.}
\usage{
pop_colors
}
\description{
List of default pop colors
}
\details{
Convenience list of pop colors in #RGB format.  Just type
names(pop_colors) to see the list of colors included.
}
\examples{

\dontrun{ggplot() + geom_line(aes(x=-5:5, y=(-5:5)^2), colour=pop_colors$gold)}

}
\keyword{datasets}
